Some more renaming of variables, just to make it more clear for myself.
Added the functionality to send a message to a target when you buff them with an aura. For instance Innervate for Druids.
The option is disabled for buffs only usable on your self. It will not send a whisper to yourself either. And it only works for players in your party/raid.

The following spells are added as well.
Combat rezz and bash for druids.
Tranq shot for hunters.

Signed-off-by: p1tcode <perivar@gmail.com>